requirements.txt
# Biblioth�que pour la gestion YAML (configuration projet)
PyYAML>=6.0

# types
types-PyYAML
types-requests

# Biblioth�ques principales Data Science et Machine Learning

# Manipulation des donn�es
pandas>=1.5.0
pandas-stubs

# Calcul num�rique
numpy>=1.23.0

# Machine Learning
scikit-learn>=1.2.0

# Visualisation
matplotlib>=3.7.0
seaborn>=0.12.0

# web crawling/scraping
bs4
playwright
